refactor(mrml-wasm): split options in modules
